+++ packages/realm_dart/lib/src/session.dart
@@ -73,15 +73,7 @@ class SyncProgress {
   /// value may either increase or decrease as new data needs to be transferred.
   final double progressEstimate;
 
-  const SyncProgress._({required this.progressEstimate});
-
-  static double _calculateProgress({required int transferred, required int transferable}) {
-    if (transferable == 0 || transferred > transferable) {
-      return 1;
-    }
-
-    return transferred / transferable;
-  }
+  const SyncProgress({required this.progressEstimate});
 }
 
 /// A type containing information about the transition of a connection state from one value to another.
@@ -108,12 +100,11 @@ extension SessionInternal on Session {
 
   void raiseError(int errorCode, bool isFatal) => handle.raiseError(errorCode, isFatal);
 
-  static SyncProgress createSyncProgress(int transferredBytes, int transferableBytes) =>
-      SyncProgress._(progressEstimate: SyncProgress._calculateProgress(transferred: transferredBytes, transferable: transferableBytes));
+  static SyncProgress createSyncProgress(double progressEstimate) => SyncProgress(progressEstimate: progressEstimate);
 }
 
 abstract interface class ProgressNotificationsController {
-  void onProgress(int transferredBytes, int transferableBytes);
+  void onProgress(double progressEstimate);
 }
 
 /// @nodoc
@@ -133,10 +124,10 @@ class SessionProgressNotificationsController implements ProgressNotificationsCon
   }
 
   @override
-  void onProgress(int transferredBytes, int transferableBytes) {
-    _streamController.add(SyncProgress._(progressEstimate: SyncProgress._calculateProgress(transferred: transferredBytes, transferable: transferableBytes)));
+  void onProgress(double progressEstimate) {
+    _streamController.add(SyncProgress(progressEstimate: progressEstimate));
 
-    if (transferredBytes >= transferableBytes && _mode == ProgressMode.forCurrentlyOutstandingWork) {
+    if (progressEstimate >= 1.0 && _mode == ProgressMode.forCurrentlyOutstandingWork) {
       _streamController.close();
     }
   }